LayoutTests/imported/w3c/web-platform-tests/css/css-writing-modes/wm-propagation-body-scroll-offset-vertical-rl-expected.txt


PASS Check initial scroll position of viewport.
FAIL Scroll to scrollLeft = -1000 should be possible. assert_equals: scrollLeft should be -1000. expected -1000 but got 0
FAIL Scroll to scrollLeft = 1000 should not be possible. assert_equals: scrollLeft should be 0. expected 0 but got 1000
